Split key into key/{max-age,refresh,reset,send,set}
