+++ server/lib/people/update.js
@@ -2,7 +2,7 @@ const config = require('@kansa/common/config')
 const { InputError } = require('@kansa/common/errors')
 const LogEntry = require('@kansa/common/log-entry')
 const { sendMail, updateMailRecipient } = require('@kansa/common/mail')
-const { setKey } = require('../key')
+const setKey = require('../key/set')
 const Person = require('./person')
 
 module.exports = updatePerson
