+++ server/lib/purchase.js
@@ -2,7 +2,7 @@ const config = require('@kansa/common/config')
 const { InputError } = require('@kansa/common/errors')
 const { sendMail } = require('@kansa/common/mail')
 const Payment = require('./types/payment')
-const { refreshKey } = require('./key')
+const refreshKey = require('./key/refresh')
 const addPerson = require('./people/add')
 const Person = require('./people/person')
 const upgradePerson = require('./people/upgrade')
